Now that windows supports SSH commands, moving SSH methods to RemoteAccount.
